+++ CHANGELOG.md
@@ -5,6 +5,7 @@ v0.7.2
 ------
 
 Fixes:
+* Added NDEBUG to icedcoffee-mac target (fixes issue with assert in RectangleBinPack)
 * Fixed setNeedsDisplay in HostViewController so as to not issue a redraw call for each call
   of setNeedsDisplay when frameUpdateMode is set to ICFrameUpdateModeOnDemand.
 * Fixed ICScheduler to call setNeedsDisplay on animated nodes ins processAnimations:.
@@ -12,6 +13,8 @@ Fixes:
   practice the timer will never fire. It's used only to prevent the runloop from exiting.
 * Animations now work correctly when the host view controller's frameUpdateMode property is
   set to ICFrameUpdateModeOnDemand.
+* Fixed ICMouseEventDispatcher and ICHostViewControllerMac to correctly handle mouse moved
+  events when the host view controller's frame update mode is set to ICFrameUpdateModeOnDemand.
 
 v0.7.1
 ------